+++ temba/api/tests.py
@@ -234,7 +234,7 @@ class WebHookTest(TembaTest):
 
         # replace our uuid of 4 with the right thing
         actionset = ActionSet.objects.get(x=4)
-        actionset.set_actions_dict([WebhookAction(str(uuid4()), org.get_webhook_url(), legacy_format=True).as_json()])
+        actionset.set_actions_dict([WebhookAction(str(uuid4()), org.get_webhook_url()).as_json()])
         actionset.save()
 
         # run a user through this flow
@@ -265,27 +265,20 @@ class WebHookTest(TembaTest):
         prepared_request = args[0]
         self.assertIn(self.channel.org.get_webhook_url(), prepared_request.url)
 
-        data = parse_qs(prepared_request.body)
+        data = json.loads(prepared_request.body)
 
-        self.assertEqual(data['channel'], [str(self.channel.id)])
-        self.assertEqual(data['channel_uuid'], [self.channel.uuid])
-        self.assertEqual(data['step'], [actionset.uuid])
-        self.assertEqual(data['text'], ["Mauve"])
-        self.assertEqual(data['attachments'], ["http://s3.com/text.jpg", "http://s3.com/text.mp4"])
-        self.assertEqual(data['flow'], [str(flow.id)])
-        self.assertEqual(data['flow_uuid'], [flow.uuid])
-        self.assertEqual(data['contact'], [self.joe.uuid])
-        self.assertEqual(data['contact_name'], [self.joe.name])
-        self.assertEqual(data['urn'], [six.text_type(self.joe.get_urn('tel'))])
-
-        values = json.loads(data['values'][0])
-
-        self.assertEqual(values[0]['category'], 'Other')
-        self.assertEqual(values[0]['category_localized'], 'Other')
-        self.assertEqual(values[0]['label'], 'color')
-        self.assertEqual(values[0]['text'], 'Mauve')
-        self.assertTrue(values[0]['time'])
-        self.assertTrue(data['time'])
+        self.assertEqual(data['channel'], {'uuid': str(self.channel.uuid), 'name': self.channel.name})
+        self.assertEqual(data['contact'], {'uuid': str(self.joe.uuid), 'name': self.joe.name, 'urn': six.text_type(self.joe.get_urn('tel'))})
+        self.assertEqual(data['flow'], {'uuid': str(flow.uuid), 'name': flow.name})
+        self.assertEqual(data['input'], {
+            'urn': 'tel:+250788123123',
+            'text': "Mauve",
+            'attachments': ["image/jpeg:http://s3.com/text.jpg", "audio/mp4:http://s3.com/text.mp4"]
+        })
+        self.assertEqual(data['results']['color']['category'], 'Other')
+        self.assertEqual(data['results']['color']['name'], 'color')
+        self.assertEqual(data['results']['color']['value'], 'Mauve')
+        self.assertEqual(data['results']['color']['input'], 'Mauve')
 
     @patch('temba.api.models.time.time')
     def test_webhook_result_timing(self, mock_time):