+++ temba/channels/handlers.py
@@ -135,7 +135,7 @@ class ZenviaHandler(CourierHandler):
 
 class MtargetHandler(CourierHandler):
     channel_name = "Mtarget"
-    courier_url = r'^mt/(?P<uuid>[a-z0-9\-]+)/(?P<action>status|receive|stop)$'
+    courier_url = r'^mt/(?P<uuid>[a-z0-9\-]+)/(?P<action>status|receive)$'
     courier_name = 'courier.mt'